@@ -9,9 +9,9 @@
     "email": "raul.mihaila@gmail.com"
   },
   "peerDependencies": {
-    "crizmas-utils": "^1.0.1",
-    "prop-types": "^15.5.10",
-    "react": "^15.5.4 || ^16.0.0"
+    "crizmas-utils": "^1.0.3",
+    "prop-types": "^15.6.2",
+    "react": "^16.5.2"
   },
   "homepage": "https://raulsebastianmihaila.github.io/crizmas-mvc-docs",
   "repository": {
